It Stares Back wipe
Working on the Inventory System currently. This push is from all of the work I did yesterday, so the framework (sorta) of the item pick ups, and now working on storage and UI when the player picks an item up. Kinda unsure as to whether there should be a menu the player accesses.